+++ test/fixtures/src/inflater_spec.dart
@@ -1,3 +1,4 @@
+import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 
 // Define constants for all widgets and objects we want to generate inflaters for.
@@ -6,8 +7,11 @@ import 'package:flutter/material.dart';
 // ignore_for_file: unused_element
 
 const AppBar? _appBar = null;
+const BoxConstraints? _boxConstraints = null;
 const Center? _center = null;
+const CircularProgressIndicator? _circularProgressIndicator = null;
 const Column? _column = null;
+const CupertinoTextField? _cupertinoTextField = null;
 const FloatingActionButton? _floatingActionButton = null;
 const Icon? _icon = null;
 const MaterialApp? _materialApp = null;
@@ -18,8 +22,11 @@ const ThemeData? _themeData = null;
 
 const inflaters = [
   AppBar,
+  BoxConstraints,
   Center,
+  CircularProgressIndicator,
   Column,
+  CupertinoTextField,
   FloatingActionButton,
   Icon,
   MaterialApp,